treewide: Support selecting different architecture
By default the PDK downloads and runs QEMU images based on
the host's CPU. For those using the PDK on Intel x86_64 it
is now possible to override the default selected architecture
via '-a=' or '--arch=' to download and run arm64 images
on Intel x86_64 machines. QEMU arguments whether to use KVM
in those situations or not have been adapted too.