scripts: Refine setup process
The Snap version of the PDK doesn't need sshd set up,
so only run those paths on non-Snap systems.

Also, on Snap systems, check for the required permissions
during setup, so that users don't have to complain about
an incomplete environment. That includes checking for
the 'removable-media' plug when offloading the
DATA_ROOT to an external medium.